refactor: Session 하위로 이미지 도메인 패키지 이동
